Revert "WaE: comparison of integers of different signs"
This is pointless, the GCC warning was fixed in the previous commit and
FT_Bitmap.width is of type "unsigned int" already.

This reverts commit bc4a2d7ce9d4417f9d6cd58505d4ea3e959b1354.

+++ vcl/generic/glyphs/gcach_ftyp.cxx
@@ -1206,7 +1206,7 @@ bool ServerFont::GetGlyphBitmap8( sal_GlyphId aGlyphId, RawBitmap& rRawBitmap ) 
         unsigned int x;
         for( int y = rRawBitmap.mnHeight; --y >= 0 ; )
         {
-            for( x = 0; x < static_cast<unsigned int>(rBitmapFT.width); ++x )
+            for( x = 0; x < rBitmapFT.width; ++x )
                 *(pDest++) = *(pSrc++);
             for(; x < rRawBitmap.mnScanlineSize; ++x )
                 *(pDest++) = 0;
@@ -1218,7 +1218,7 @@ bool ServerFont::GetGlyphBitmap8( sal_GlyphId aGlyphId, RawBitmap& rRawBitmap ) 
         for( int y = rRawBitmap.mnHeight; --y >= 0 ; )
         {
             unsigned char nSrc = 0;
-            for( x = 0; x < static_cast<unsigned int>(rBitmapFT.width); ++x, nSrc+=nSrc )
+            for( x = 0; x < rBitmapFT.width; ++x, nSrc+=nSrc )
             {
                 if( (x & 7) == 0 )
                     nSrc = *(pSrc++);
